FIXED: Vulnerability ID: HTB22920 Reference: http://www.htbridge.ch/advisory/path_disclosure_in_viscacha.html Product: Viscacha Vendor: MaMo Net ( http://www.viscacha.org ) Vulnerable Version: 0.8.1 Vulnerability Type: Path disclosure Risk level: Low Credit: High-Tech Bridge SA - Ethical Hacking & Penetration Testing (http://www.htbridge.ch/)
Vulnerability Details:
The vulnerability exists due to failure in the "editprofile.php" script, it's possible to generate an error that will reveal the full path of the script.
A remote user can determine the full path to the web root directory and other potentially sensitive information.
 
The following PoC is available:
http://[host]/editprofile.php?action=addabo
http://[host]/editprofile.php?action=removeabo